+++ src/main/java/tconstruct/library/armor/BehaviorDispenseArmorCopy.java
@@ -21,7 +21,7 @@ final class BehaviorDispenseArmorCopy extends BehaviorDefaultDispenseItem {
         int j = par1IBlockSource.getYInt() + enumfacing.getFrontOffsetY();
         int k = par1IBlockSource.getZInt() + enumfacing.getFrontOffsetZ();
         AxisAlignedBB axisalignedbb = AxisAlignedBB.getBoundingBox(
-                (double) i, (double) j, (double) k, (double) (i + 1), (double) (j + 1), (double) (k + 1));
+                i, j, k, i + 1, j + 1, k + 1);
         List list = par1IBlockSource
                 .getWorld()
                 .selectEntitiesWithinAABB(
